requirements.txt
beautifulsoup4
bs4
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
gunicorn==19.9.0
idna==2.7
Pillow
pytz==2018.5
requests==2.20.0
urllib3==1.24
virtualenv==16.0.0
whitenoise==4.1.2
